+++ src/JabbR-Core/Views/Account/index.cshtml
@@ -35,7 +35,7 @@
                     <div class="span12">
                         <div class="user-details">
                             <img class="gravatar" src="@gravatar" /> 
-                            <span class="name">Name</span> <!--atsign userModel.name-->
+                            <span class="name">@userModel.Name</span> <!--atsign userModel.name-->
                         </div>
                         <hr />
                     </div>
@@ -170,7 +170,7 @@
                             </form>
                         </div>
                         <div class="account-details tab-pane hide" id="identityProviders">
-                            @Html.Partial("_social")<!--, atsign userModel.SocialDetails)-->
+                            @Html.Partial("_social", @userModel.SocialDetails)
                         </div>
                     </div>
                 </div>
