+++ src/JabbR-Core/Views/Account/register.cshtml
@@ -1,11 +1,12 @@
-﻿@using JabbR;
+﻿@using JabbR_Core;
+
 @{
     Layout = "_layout.cshtml";
 }
 <div class="row-fluid">
     <div class="span6 offset3">
         <div class="user-registration">
-            <form class="form-horizontal" action="@Url.Content("~/account/create" + Html.RequestQuery())" method="post">
+            <form class="form-horizontal" action="@Url.Content("~/account/create")" method="post">  <!--/* + Html.RequestQuery()*/-->
                 @Html.ValidationSummary()
                 <div class="control-group">
                     <div class="controls">
@@ -23,8 +24,8 @@
                         </div>
                     </div>
                 </div>
-                @if (ViewBag.requirePassword) 
-                {
+          <!--      atsignif (ViewBag.requirePassword) 
+                {-->
                 <div class="control-group">
                     <div class="controls">
                         <div class="input-prepend">
@@ -41,11 +42,11 @@
                         </div>
                     </div>
                 </div>
-                }
+              <!--  } -->
                 <div class="control-group">
                     <div class="controls">
                         <button type="submit" class="btn">@LanguageResources.Account_Register</button>
-                        <a href="@Url.Content("~/account/login" + Html.RequestQuery())" class="btn">@LanguageResources.Account_Cancel</a>
+                        <a href="@Url.Content("~/account/login" /* + Html.RequestQuery()*/)" class="btn">@LanguageResources.Account_Cancel</a>
                     </div>
                 </div>
                 @Html.AntiForgeryToken()
