+++ src/JabbR-Core/Views/Account/requestresetpassword.cshtml
@@ -5,7 +5,7 @@
 <div class="row-fluid">
     <div class="span6 offset3">
         <div class="user-request-reset-password">
-      <!--      <form class="form-horizontal" action="atsign Url.Content("~/account/requestresetpassword" + Html.RequestQuery())" method="post">
+           <form class="form-horizontal" action="@Url.Content("~/account/requestresetpassword" /* + Html.RequestQuery()*/)" method="post">
                 @Html.ValidationSummary()
                 <div class="control-group">
                     <div class="controls">
@@ -18,7 +18,7 @@
                 <div class="control-group">
                     <div class="controls">
                         <button type="submit" class="btn">@LanguageResources.Account_ResetRequest</button>
-                        <a href="atsign Url.Content("~/account/login" + Html.RequestQuery())" class="btn">@LanguageResources.Account_Cancel</a>
+                        <a href="@Url.Content("~/account/login" /* + Html.RequestQuery()*/)" class="btn">@LanguageResources.Account_Cancel</a>
                     </div>
                 </div>
                 @Html.AntiForgeryToken()
