+++ src/JabbR-Core/Views/Account/requestresetpasswordsuccess.cshtml
@@ -5,7 +5,7 @@
 <div class="row-fluid">
     <div class="span12">
         <div class="user-reset-password">
-            @Html.Raw(String.Format(LanguageResources.Account_ResetPasswordSuccessMessage, @Model))
+            @Html.Raw(@LanguageResources.Account_ResetPasswordSuccessMessage)<!--/*, atsignModel*/))-->
         </div>
     </div>
 </div>
