+++ src/JabbR-Core/Views/Account/resetpassword.cshtml
@@ -5,7 +5,7 @@
 <div class="row-fluid">
     <div class="span6 offset3">
         <div class="user-request-reset-password">
-            <form class="form-horizontal" action="@Url.Content("~/account/resetpassword/" + Model /*+ Html.RequestQuery()*/)" method="post">
+            <form class="form-horizontal" action="@Url.Content("~/account/resetpassword/" + Model /* + Html.RequestQuery()*/)" method="post">
                 @Html.ValidationSummary()
                 <div class="control-group">
                     <div class="controls">
