+++ src/JabbR-Core/Views/Account/resetpassworderror.cshtml
@@ -1,4 +1,4 @@
-﻿@using JabbR;
+﻿@using JabbR_Core;
 @{
     Layout = "_layout.cshtml";
 }
